requirements.txt
requests
beautifulsoup4
pandas
numpy
sqlalchemy
pandastable
Pillow
matplotlib
scikit-learn